requirements.txt
pytest
pytest-pylint
python-dotenv
ratelimiter